pyproject.toml
# MPlus 项目配置文件
# AI驱动的自媒体选题助手

[tool.poetry]
name = "mplus"
version = "0.1.0"
description = "AI驱动的自媒体选题助手，通过头脑风暴优化选题，并用社交模拟预测内容效果"
authors = ["MPlus Team"]
readme = "README.md"
# 指定实际的包位置
packages = [
    { include = "backend" },
    { include = "tui" },
]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"

# Web 框架
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
websockets = "^12.0"

# AI 框架（Phase 2 使用）
# 注意：使用基础版本避免过多间接依赖，后续按需添加 extras
camel-ai = "^0.2.0"

# 模拟引擎 — Ripple CAS 社会行为预测引擎
# 开发阶段使用本地路径，生产部署可切换为 git 依赖：
# ripple = {git = "https://github.com/xyskywalker/Ripple.git", tag = "v0.2.0"}
ripple = {path = "../Ripple", develop = true}

# 数据库
sqlalchemy = "^2.0.0"
aiosqlite = "^0.19.0"

# HTTP 客户端
httpx = "^0.27.0"

# 配置管理
pydantic = "^2.0.0"
pydantic-settings = "^2.0.0"
python-dotenv = "^1.0.0"

# TUI
rich = "^13.0.0"

# 工具
pyyaml = "^6.0"
cachetools = "^5.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
ruff = "^0.1.0"

[tool.poetry.scripts]
mplus = "tui.main:main"

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
